+++ cms-picture-gallery/src/app/services/core-content.service.ts
@@ -1,33 +1,23 @@
 import { Injectable } from '@angular/core';
-import { TreeNode, TreeNodeCollection, TopLevelCollection, ShreeGaneshCollection } from 'src/assets/content-tree/tree-nodes' ;
+import { TreeNodeElement,MenuTreeElements } from 'src/assets/content-tree/tree-nodes' ;
 import { BreadCrumb } from 'src/assets/content-tree/bread-crumbs';
-import { parseTemplate } from '@angular/compiler';
 @Injectable({
   providedIn: 'root'
 })
 export class CoreContentService {
 
   constructor() { }
-  private breadCrumbs:BreadCrumb[] =[] ;
+  private breadCrumbs?:BreadCrumb[] =[] ;
   
-  get BreadCrumbs():BreadCrumb[] {
+  get BreadCrumbs():BreadCrumb[]| undefined{
     return this.breadCrumbs;
   } 
-  setCurrentMenu():TreeNodeCollection  {
-    let retElement:TreeNodeCollection = TopLevelCollection;
-    let currentMenu =  localStorage.getItem('current-menu') ;
-    let parentKey = null;
-    switch(currentMenu) {
-      case 'top-level': 
-      retElement =  TopLevelCollection ;
-        break ;
-      case 'religious-shree-ganesh':
-        retElement =  ShreeGaneshCollection ;
-      
-          break ;
-    }
-    this.breadCrumbs  = retElement.breadCrumb;
-    return retElement ;
+  setCurrentCardList():TreeNodeElement[]  {
+    let retCardList:TreeNodeElement[] = [];
+    let currentParentKey =  localStorage.getItem('current-menu') ;
+    retCardList = MenuTreeElements.filter((x) => x.parentKey === currentParentKey) ;
+     
+    this.breadCrumbs  = MenuTreeElements.find((x) => x.key === currentParentKey)?.breadCrumb ;
+    return retCardList ;
   }
- 
 }