+++ cms-picture-gallery/src/assets/content-tree/tree-nodes.ts
@@ -6,166 +6,201 @@ import { BreadCrumb } from "./bread-crumbs";
 export interface TreeNodeElement {
     key:string; // unique id.. like a primary key, also picture list loader for leaf
     parentKey?:string; // like a foreign key and establishing parent. optional()
-    levelIndex?:number ;
+    levelIndex?:number ; // unused for now (May 2024)
     breadCrumb:BreadCrumb[]; // better safe than optimal for now May 3 2024
-    // better safe than optimal 
     label:string; 
     description?:string ;
     roleBasedDescription?: string;
-    isLeaf:boolean;
+    isCompiledList?:boolean; // use some criteria to load selected images like latest uploads or top rated
+    isLeafParent:boolean; // 'true' = load the image list for that menu kee, 'false' = top or middle node
     accessible?:string[]; // roles match users roles many to many match
     
 }
 
+export const MenuTreeElements:TreeNodeElement[] = [
+    {
+        key:'top-level', levelIndex:0,isLeafParent:false,accessible:["all"], // roles match users roles many to many match
+        breadCrumb:[
+            {link: '', params:'top-level', label: 'HOME'}
+        ],
+        label:"HOME PAGE",
+    },
+    {
+        parentKey:'top-level',key: "special-lists", levelIndex:1,isLeafParent:false, accessible:['all'],
+        label: "Special Lists", 
+        description:'Extracted works from the entire collection like best attempts, something new, latest sketches, etc',
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '', params:'special-lists', label: 'Special Lists'}
+        ],    
+    }, 
+    { 
+        parentKey: "special-lists", key: 'changers', levelIndex:2, isLeafParent: false, isCompiledList:true, accessible: ["all"], 
+        label: 'Changers', 
+        description:`These are landmark sketches which I consider a significant change or turn in the progress of my sketches, or maybe a special reason. 
+                    These may not be my best efforts but are a new element or entity that was introduced in these drawings.`,
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'special-lists', label: 'Special Lists'},
+            {link: '', params:'changers', label: 'Changers'}
+        ],
+    }, 
+    { 
+        parentKey: "changers", key: 'changers-b4-2022', levelIndex:3, isLeafParent: true, isCompiledList:true, accessible: ["all"], 
+        label: 'Changers: before 2022', 
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'special-lists', label: 'Special Lists'},
+            {link: '/view', params:'changers', label: 'Changers'},
+            {link: '', params:'changers-b4-2022', label: 'Changers: before 2022'}
+        ],
+    },
+    { 
+        parentKey: "changers", key: 'changers-2022', levelIndex:3,  isLeafParent: true, isCompiledList:true, accessible: ["all"],
+        label:  'Changers: 2022',
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'special-lists', label: 'Special Lists'},
+            {link: '/view', params:'changers', label: 'Changers'},
+            {link: '', params:'changers-2022', label: 'Changers: 2022'}
+        ], 
+    },
+    { 
+        parentKey: "changers",key: 'changers-2023', levelIndex:3,  isLeafParent: true,  accessible: ["all"],
+        label:  'Changers: 2023',  
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'special-lists', label: 'Special Lists'},
+            {link: '/view', params:'changers', label: 'Changers'},
+            {link: '', params:'changers-2023', label: 'Changers: 2023'}
+        ], 
+    },
+    { 
+        parentKey: "changers",key: 'changers-2024', levelIndex:3, isLeafParent: true,  accessible: ["all"],
+        label:  'Changers: 2024', 
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'special-lists', label: 'Special Lists'},
+            {link: '/view', params:'changers', label: 'Changers'},
+            {link: '', params:'changers-2024', label: 'Changers: 2024'}
+        ], 
+    },
+    {
+        parentKey:'top-level', key:"religion-and-gurus",  levelIndex:1, isLeafParent: false,  accessible: ["sanatan", "superuser","guru"],
+        label:  'Religion And Gurus',
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '', params:'religion-and-gurus', label: 'Religion And Gurus'},
+        ], 
+    },
+    {
+        parentKey:"religion-and-gurus", key: "religious-shree-ganesh", levelIndex:2,isLeafParent:false, accessible:['superuser','sanatan'],
+        label: "Religious: Shree Ganesh",
+        description:'Sketches of, and including Shree Ganesh',
+        
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'religion-and-gurus', label: 'Religion And Gurus'},
+            {link: '', params:'religious-shree-ganesh', label: 'Shree Ganesh'}
+        ],
+        
+    },
+    {
+        parentKey:"religious-shree-ganesh", key: "shree-ganesh-b4-q4-2021", levelIndex:3, isLeafParent:true, accessible:['superuser','sanatan'],
+        label: "Shree Ganesh Before Q4 2021",
+        description:'Sketches of, and including Shree Ganesh prior to Q4 2021', 
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'religion-and-gurus', label: 'Religion And Gurus'},
+            {link: '/view', params:'religious-shree-ganesh', label: 'Shree Ganesh'},
+            {link: '', params:'shree-ganesh-b4-q4-2021', label: 'Shree Ganesh Before Q4 2021'}
+        ],
+        
+    },
+    {
+        parentKey:"religious-shree-ganesh", key: "shree-ganesh-gte-q1-2023", levelIndex:3, isLeafParent:true, accessible:['superuser','sanatan'],
+        label: "Shree Ganesh Q1 2023 onward",
+        description:'Sketches of, and including Shree Ganesh Q1 2023 to Q1 2024', 
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'religion-and-gurus', label: 'Religion And Gurus'},
+            {link: '/view', params:'religious-shree-ganesh', label: 'Shree Ganesh'},
+            {link: '', params:'shree-ganesh-gte-q1-2023', label: 'Shree Ganesh Q1 2023 onward'}
+        ],
+        
+    },
+    {
+        parentKey:"religious-shree-ganesh", key: "shree-ganesh-gte-q1-2024", levelIndex:3, isLeafParent:true, accessible:['superuser','sanatan'],
+        label: "Shree Ganesh Q1 2024 onward",
+        description:'Sketches of, and including Shree Ganesh Q1 2023 onward', 
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'religion-and-gurus', label: 'Religion And Gurus'},
+            {link: '/view', params:'religious-shree-ganesh', label: 'Shree Ganesh'},
+            {link: '', params:'shree-ganesh-b4-q4-2021', label: 'Shree Ganesh Q1 2024 onward'}
+        ],        
+    },
+    {
+        parentKey:"religion-and-gurus", key: "devi", levelIndex:2,isLeafParent:false, accessible:['superuser','sanatan'],
+        label: "Religious: Goddesses",
+        description:'Sketches of, and including Shree Ganesh',
+        
+        breadCrumb:[
+            {link: '/view', params:'top-level', label: 'HOME'},
+            {link: '/view', params:'religion-and-gurus', label: 'Religion And Gurus'},
+            {link: '', params:'religious-shree-ganesh', label: 'Shree Ganesh'}
+        ],
+        
+    },  
+]
 /*
 
 export const MenuTree: MenuItem[] = [
-    { tab:"Admin", label:  'Technical', isLeaf: false, key: 'technical', roles: ['non-living', 'all'], dateUploaded: '05-12-2021' },
-    { parentKey: "special-lists", label: 'Changers: before 2022', isLeaf: false, key: '',  key: 'changers-b4-2022', roles: ["any"] },
-    { parentKey: "special-lists", label:  ' - 2022', isLeaf: true, key: 'changers-2022', roles: ["any"] },
-    { parentKey: "special-lists", label:  ' - 2023', isLeaf: true, key: 'changers-2023', roles: ["any"] },
-    { parentKey: "special-lists", label:  ' - 2024', isLeaf: true, key: 'changers-2024', roles: ["any"] },
-    { parentKey: "special-lists", label:  'The Best: before 2022', isLeaf: false, key: 'showpiece', roles: ["any"] },
-    { parentKey: "special-lists", label:  '- 2022', isLeaf: false, key: 'showpiece-2022', roles: ["any"] },
-    { parentKey: "special-lists", label:  '- 2023', isLeaf: false, key: 'showpiece-2023', roles: ["any"] },
-    { parentKey: "special-lists", label:  '- 2024', isLeaf: false, key: 'showpiece-2024', roles: ["any"] },
-    { parentKey: "special-lists", label:  'Latest Uploads', isLeaf: false, key: 'latest-uploads', roles: ['any'] },
-    { parentKey: "religious-shree-ganesh", roles:["sanatani"], label:  'Shree Ganesh: pre Q42021 ', isLeaf: false, ey: 'shree-ganesh',  dateUploaded: '12-11-2021' },
-    { parentKey: "religious-shree-ganesh", roles:["sanatani"], label:  '- Q4 2021 onward', isLeaf: true, key: 'shree-ganesh-gte-q4-2021',  dateUploaded: '01-31-2023' },
-    { parentKey: "religious-shree-ganesh", roles:["sanatani"], label:  '- Q1 2023 onward', isLeaf: true, key: 'shree-ganesh-gte-q1-2023',  dateUploaded: '03-23-2023' },
-    { parentKey: "religious-shree-ganesh", roles:["sanatani"], label:  '- Q1 2024 onward', isLeaf: true, key: 'shree-ganesh-gte-q1-2024',  dateUploaded: '04-09-2024' },
-    { parentKey: "religious-other-deities", label:  'Goddesses', isLeaf: false, key: 'devi', roles: ["sanatani"], dateUploaded: '01-29-2023' },
-    { parentKey: "religious-other-deities", label:  'Mahadev', isLeaf: false, key: 'mahadev', roles: ["sanatani"], dateUploaded: '02-01-2023' },
-    { parentKey: "religious-other-deities", label:  'Mahadev and Family', isLeaf: false, key: 'mahadev-family', roles: ["sanatani"], dateUploaded: '06-01-2023' },
-    { parentKey: "religious-other-deities", label:  'Laxmi, Narayan, Shree Ram, and family, Bajrangbali', isLeaf: false,  dateUploaded: '02-05-2023', key: 'laxmi-vishnu-hanuman', roles: ["sanatani"] },
-    { parentKey: "religious-gurus-and-mystics", label:  'Spiritual Souls:Datta Lineage, Sikhs, East of India', isLeaf: false, parent:true, dateUploaded: '02-13-2023', key: 'dattavatar', roles: ["guru"] },
-    { parentKey: "religious-gurus-and-mystics", label:  'Shree Swami Samartha', isLeaf: false, key: 'swami-samartha', roles: ["sanatani"], dateUploaded: '04-24-2023' },
-    { parentKey: "religious-gurus-and-mystics", label:  'Shree Swami Samartha (Q2 2023 onward)', isLeaf: false, key: 'swami-samartha-q2-2023', roles: ["guru"], dateUploaded: '04-13-2024' },
-    { parentKey: "shirdi-sai", label:  'Shirdi Sai: to  Q3 2021', isLeaf: false, parent: true, key: 'shirdi-sai-q1-q2-2021', roles: ["guru"], dateUploaded: '02-07-2023'  },
-    { parentKey: "shirdi-sai", label:  '- Q3 2021 - Q1 2022', isLeaf: true, key: 'shirdi-sai-q3-q4-2021', roles: ["guru"], dateUploaded: '03-31-2022' },
-    { parentKey: "shirdi-sai", label:  '- Q2 Q3 2022', isLeaf: true, key: 'shirdi-sai-q2-q3-2022', roles: ["guru"], dateUploaded: '02-07-2023' },
-    { parentKey: "shirdi-sai", label:  '- Q3,Q4 2022 Q1 2023', isLeaf: true, key: 'shirdi-sai-q4-2022-q1-2023', roles: ["guru"], dateUploaded: '03-02-2023' },
-    { parentKey: "shirdi-sai", label:  '- Q2 Q3 2023', isLeaf: true, key: 'shirdi-sai-q2-q3-2023', roles: ["guru"], dateUploaded: '06-02-2023' },
-    { parentKey: "shirdi-sai", label:  '- Q4 2023 Q1 2024', isLeaf: true, key: 'shirdi-sai-q4-2023-q1-2024', roles: ["guru"], dateUploaded: '04-14-2024' },
-    { parentKey: "shirdi-sai", label:  '- Sai Satcharitra', isLeaf: true, key: 'baba-themes-1', roles: ["guru"], dateUploaded: '02-28-2022' },
-    { parentKey: "otherbeings-places-events", label:  'Religion and Gurus', isLeaf: false, key: 'religion-and-guru', roles: ['people',  'non-religious', 'all'], dateUploaded: '02-02-2023' },
-     { parentKey: "otherbeings-places-events", label:  'People', isLeaf: false, key: 'people-places', roles: ['people',  'non-religious', 'all'], dateUploaded: '04-14-2024' },
-    { parentKey: "otherbeings-places-events", label:  'Places, Scenes and Objects', isLeaf: false, key: 'places-scenes-objects', roles: ['non-living',  'non-religious', 'all'], dateUploaded: '06-01-2023' },
-    { parentKey: "otherbeings-places-events", label:  'Misc, Themes', isLeaf: false, key: 'themes-misc', roles: ['non-living',  'non-religious', 'all'], dateUploaded: '05-14-2023' },
-    { tab:"Misc", label:  'Anim Unorganized', isLeaf: false, key: 'animate-to-be-oragnized1', roles: ['non-living',  'non-religious', 'all'], dateUploaded: '02-03-2022' },
-   // WIP { tab:"Misc", label:  'Technical', isLeaf: false, key: 'technical', roles: ['non-living', 'all'], dateUploaded: '02-03-2022' },
-    { parentKey: "transport-and-machines", label:  'Rail', isLeaf: false,  key: 'trains', roles: ['non-living', 'non-religious',  'all'], dateUploaded: '02-08-2023' },
-    { parentKey: "transport-and-machines", label:  ' - IR WDM2', isLeaf: true, key: 'trains-ir-alcos', roles: ['non-living', 'non-religious',  'all'], dateUploaded: '11-28-2022' },
-    { parentKey: "transport-and-machines", label:  ' - IR trains)', isLeaf: true, key: 'trains-ir-special-trains', roles: ['non-living',  'non-religious', 'all'], dateUploaded: '02-02-2023' },
-    { parentKey: "mumbai-meri-jaan", label:  'Mumbai Meri Jaan', isLeaf: false, key: 'mumbai-meri-jaan', roles: ['non-living',  'non-religious', 'all'], dateUploaded: '11-28-2022' },
-    { parentKey: "mumbai-meri-jaan", label:  ' - Q1 2023 onward', isLeaf: true, key: 'mumbai-meri-jaan-2', roles: ['non-living',  'non-religious', 'all'], dateUploaded: '01-29-2023' },
-    { parentKey: "transport-and-machines", label:  'Aircraft, Ships and Cars', isLeaf: false, key: 'planes', roles: ['non-living', 'non-religious',  'all'], dateUploaded: '02-14-2023' },
-    { parentKey: "transport-and-machines", label:  '- Q2 2023 onward', isLeaf: false, key: 'planesQ12023', roles: ['non-living', 'non-religious', 'all'], dateUploaded: '05-12-2023' },
-    { tab:"Admin", label:  'technical', isLeaf: false, key: 'technical', roles: ['non-living', 'all'], dateUploaded: '05-12-2021' },
+    { tab:"Admin", label:  'Technical', isLeafParent: false, key: 'technical', accessible: ['non-living', 'all'], dateUploaded: '05-12-2021' },
+   
+    
+   
+    { parentKey: "special-lists", label:  'The Best: before 2022', isLeafParent: false, key: 'showpiece', accessible: ["all"] },
+    { parentKey: "special-lists", label:  '- 2022', isLeafParent: false, key: 'showpiece-2022', accessible: ["all"] },
+    { parentKey: "special-lists", label:  '- 2023', isLeafParent: false, key: 'showpiece-2023', accessible: ["all"] },
+    { parentKey: "special-lists", label:  '- 2024', isLeafParent: false, key: 'showpiece-2024', accessible: ["all"] },
+    { parentKey: "special-lists", label:  'Latest Uploads', isLeafParent: false, key: 'latest-uploads', accessible: ['all'] },
+    { parentKey: "religious-shree-ganesh", accessible:["sanatani"], label:  'Shree Ganesh: pre Q42021 ', isLeafParent: false, ey: 'shree-ganesh',  dateUploaded: '12-11-2021' },
+    { parentKey: "religious-shree-ganesh", accessible:["sanatani"], label:  '- Q4 2021 onward', isLeafParent: true, key: 'shree-ganesh-gte-q4-2021',  dateUploaded: '01-31-2023' },
+    { parentKey: "religious-shree-ganesh", accessible:["sanatani"], label:  '- Q1 2023 onward', isLeafParent: true, key: 'shree-ganesh-gte-q1-2023',  dateUploaded: '03-23-2023' },
+    { parentKey: "religious-shree-ganesh", accessible:["sanatani"], label:  '- Q1 2024 onward', isLeafParent: true, key: 'shree-ganesh-gte-q1-2024',  dateUploaded: '04-09-2024' },
+    { parentKey: "religious-other-deities", label:  'Goddesses', isLeafParent: false, key: 'devi', accessible: ["sanatani"], dateUploaded: '01-29-2023' },
+    { parentKey: "religious-other-deities", label:  'Mahadev', isLeafParent: false, key: 'mahadev', accessible: ["sanatani"], dateUploaded: '02-01-2023' },
+    { parentKey: "religious-other-deities", label:  'Mahadev and Family', isLeafParent: false, key: 'mahadev-family', accessible: ["sanatani"], dateUploaded: '06-01-2023' },
+    { parentKey: "religious-other-deities", label:  'Laxmi, Narayan, Shree Ram, and family, Bajrangbali', isLeafParent: false,  dateUploaded: '02-05-2023', key: 'laxmi-vishnu-hanuman', accessible: ["sanatani"] },
+    { parentKey: "religious-gurus-and-mystics", label:  'Spiritual Souls:Datta Lineage, Sikhs, East of India', isLeafParent: false, parent:true, dateUploaded: '02-13-2023', key: 'dattavatar', accessible: ["guru"] },
+    { parentKey: "religious-gurus-and-mystics", label:  'Shree Swami Samartha', isLeafParent: false, key: 'swami-samartha', accessible: ["sanatani"], dateUploaded: '04-24-2023' },
+    { parentKey: "religious-gurus-and-mystics", label:  'Shree Swami Samartha (Q2 2023 onward)', isLeafParent: false, key: 'swami-samartha-q2-2023', accessible: ["guru"], dateUploaded: '04-13-2024' },
+    { parentKey: "shirdi-sai", label:  'Shirdi Sai: to  Q3 2021', isLeafParent: false, parent: true, key: 'shirdi-sai-q1-q2-2021', accessible: ["guru"], dateUploaded: '02-07-2023'  },
+    { parentKey: "shirdi-sai", label:  '- Q3 2021 - Q1 2022', isLeafParent: true, key: 'shirdi-sai-q3-q4-2021', accessible: ["guru"], dateUploaded: '03-31-2022' },
+    { parentKey: "shirdi-sai", label:  '- Q2 Q3 2022', isLeafParent: true, key: 'shirdi-sai-q2-q3-2022', accessible: ["guru"], dateUploaded: '02-07-2023' },
+    { parentKey: "shirdi-sai", label:  '- Q3,Q4 2022 Q1 2023', isLeafParent: true, key: 'shirdi-sai-q4-2022-q1-2023', accessible: ["guru"], dateUploaded: '03-02-2023' },
+    { parentKey: "shirdi-sai", label:  '- Q2 Q3 2023', isLeafParent: true, key: 'shirdi-sai-q2-q3-2023', accessible: ["guru"], dateUploaded: '06-02-2023' },
+    { parentKey: "shirdi-sai", label:  '- Q4 2023 Q1 2024', isLeafParent: true, key: 'shirdi-sai-q4-2023-q1-2024', accessible: ["guru"], dateUploaded: '04-14-2024' },
+    { parentKey: "shirdi-sai", label:  '- Sai Satcharitra', isLeafParent: true, key: 'baba-themes-1', accessible: ["guru"], dateUploaded: '02-28-2022' },
+    { parentKey: "otherbeings-places-events", label:  'Religion and Gurus', isLeafParent: false, key: 'religion-and-guru', accessible: ['people',  'non-religious', 'all'], dateUploaded: '02-02-2023' },
+     { parentKey: "otherbeings-places-events", label:  'People', isLeafParent: false, key: 'people-places', accessible: ['people',  'non-religious', 'all'], dateUploaded: '04-14-2024' },
+    { parentKey: "otherbeings-places-events", label:  'Places, Scenes and Objects', isLeafParent: false, key: 'places-scenes-objects', accessible: ['non-living',  'non-religious', 'all'], dateUploaded: '06-01-2023' },
+    { parentKey: "otherbeings-places-events", label:  'Misc, Themes', isLeafParent: false, key: 'themes-misc', accessible: ['non-living',  'non-religious', 'all'], dateUploaded: '05-14-2023' },
+    { tab:"Misc", label:  'Anim Unorganized', isLeafParent: false, key: 'animate-to-be-oragnized1', accessible: ['non-living',  'non-religious', 'all'], dateUploaded: '02-03-2022' },
+   // WIP { tab:"Misc", label:  'Technical', isLeafParent: false, key: 'technical', accessible: ['non-living', 'all'], dateUploaded: '02-03-2022' },
+    { parentKey: "transport-and-machines", label:  'Rail', isLeafParent: false,  key: 'trains', accessible: ['non-living', 'non-religious',  'all'], dateUploaded: '02-08-2023' },
+    { parentKey: "transport-and-machines", label:  ' - IR WDM2', isLeafParent: true, key: 'trains-ir-alcos', accessible: ['non-living', 'non-religious',  'all'], dateUploaded: '11-28-2022' },
+    { parentKey: "transport-and-machines", label:  ' - IR trains)', isLeafParent: true, key: 'trains-ir-special-trains', accessible: ['non-living',  'non-religious', 'all'], dateUploaded: '02-02-2023' },
+    { parentKey: "mumbai-meri-jaan", label:  'Mumbai Meri Jaan', isLeafParent: false, key: 'mumbai-meri-jaan', accessible: ['non-living',  'non-religious', 'all'], dateUploaded: '11-28-2022' },
+    { parentKey: "mumbai-meri-jaan", label:  ' - Q1 2023 onward', isLeafParent: true, key: 'mumbai-meri-jaan-2', accessible: ['non-living',  'non-religious', 'all'], dateUploaded: '01-29-2023' },
+    { parentKey: "transport-and-machines", label:  'Aircraft, Ships and Cars', isLeafParent: false, key: 'planes', accessible: ['non-living', 'non-religious',  'all'], dateUploaded: '02-14-2023' },
+    { parentKey: "transport-and-machines", label:  '- Q2 2023 onward', isLeafParent: false, key: 'planesQ12023', accessible: ['non-living', 'non-religious', 'all'], dateUploaded: '05-12-2023' },
+    { tab:"Admin", label:  'technical', isLeafParent: false, key: 'technical', accessible: ['non-living', 'all'], dateUploaded: '05-12-2021' },
 ];
 */
-export const treeNodeElements:TreeNodeElement[] = [
-    {
-        key:'top-level',
-        levelIndex:0,
-        breadCrumb:[
-            {link: '/view', params:'top-level', label: 'HOME'}
-        ],
-        label:"HOME PAGE",
-        isLeaf:false,
-        accessible:["all"] // roles match users roles many to many match
-        
-    },{
-        parentKey:'top-level',
-        key: "special-lists",
-        levelIndex:1,
-        label: "Special Lists",
-        description:'Extracted works from the entire collection like best attempts, something new, latest sketches, etc',
-        isLeaf:false,
-        accessible:['all'],
-        breadCrumb:[
-            {link: '/view', params:'top-level', label: 'HOME'},
-            {link: '/view', params:'special-lists', label: 'Special Lists'}
-        ],
-        
-    }, {
-        parentKey:'top-level',
-        key: "religious-shree-ganesh",
-        levelIndex:1,
-        label: "Religious: Shree Ganesh",
-        description:'Sketches of, and including Shree Ganesh',
-        isLeaf:false,
-        accessible:['superuser','sanatan'],
-        breadCrumb:[
-            {link: '/view', params:'top-level', label: 'HOME'},
-            {link: '/view', params:'religious-shree-ganesh', label: 'Shree Ganesh'}
-        ],
-        
-    }, 
-    /*
-
-    ,
-    {
-        key: "religious-other-deities",
-        label: "Religious: Other Deities, etc",
-        description:'Shankar Parvati, Vishnu. Other drawings in this genre',
-        isLeaf:false,
-        accessible:['superuser','sanatan']
-    }
-    ,
-    {
-        key: "religious-gurus-and-mystics",
-        label: "Religious: Gurus And Mystics",
-        description:`Gurus and Preachers related to Spirituality, like Lord Dattatreya and Avatars, the Sikh Gurus. 
-                        Shirdi Sai sketches being many are under a separate bucket`,
-        isLeaf:false,
-        accessible:['superuser','guru']
-    }
-    
-    ,
-    {
-        key: "shirdi-sai",
-        label: "Religious: Shirdi Sai Baba",
-        description:`Sadguru Shirdi Sai I have made so many drawings that it makes logical sense to put this in a separate bucket`,
-        isLeaf:false,
-        accessible:['superuser','guru']
-    }, {
-        key: "otherbeings-places-events",
-        label: "Beings, Places and Themes ",
-
-        description:`Pictures of the living - famous human personalities, other living beings (as in May 2024 occasional). Places and famous structures. Events, Misc. Themes. 
-                  (But Mumbai City related drawing warrant a separate, independent group!!)`,
-        roleBasedDescription:`ADD:Also some chosen religious and spirtual themes: NOTS superuser, sanatan,guru'`,
-        
-        // if exists rolebaseddescr
-        //    split by ':'
-        //    if user roles match third segment NOTS = not or nothing means match the role
-        //    1st segment, ADD or no spec = add, INSERT = add at begining, REPLACE = replace
-        // Rest logic here, else if will get too complicated
-        
-        isLeaf:false,
-        accessible:['superuser','people'],// ??=>, 'all']
-    }, {
-        key: "mumbai-meri-jaan",
-        label: "Salaam Mumbai",
-
-        description:`Dedication to the City that is part of me that I am part of. Wherever I am`,
-        isLeaf:false,
-        accessible:['superuser','people'],// ??=>, 'all']
-    },
-    {
-        key: "transport-and-machines",
-        label: "Transport and Machines",
-        description:`Trains, Cars, Planes, Ships, Military equipment. Anything that is mechanical`,
-        isLeaf:false,
-        accessible:['all']
-    },
-    **/
-    
-    
-]
 // #### IN TRANSIT .. OLDER ITERATION
-
+/*
 
 export interface TreeNode {
     key:string; // unique id.. like a primary key, also picture list loader for leaf
@@ -174,7 +209,7 @@ export interface TreeNode {
     label:string; 
     description?:string ;
     roleBasedDescription?: string;
-    isLeaf:boolean;
+    isLeafParent:boolean;
     accessible?:string[]; // roles match users roles many to many match
     isLeafren?:TreeNode[]; // noSQL wannabe; unused now
 }
@@ -188,20 +223,20 @@ export const TopLevelCollection:TreeNodeCollection =
 {
     levelIndex:0,
     breadCrumb:[
-        {link: '/view', params:'top-level', label: 'HOME'}
+        {link: '', params:'top-level', label: 'HOME'}
     ],
     treeNodeElements:[{
         key: "special-lists",
         label: "Special Lists",
         description:'Extracted works from the entire collection like best attempts, something new, latest sketches, etc',
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['all']
     },
     {
         key: "religious-shree-ganesh",
         label: "Religious: Shree Ganesh",
         description:'Sketches of, and including Shree Ganesh',
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['superuser','sanatan']
     }
     ,
@@ -209,7 +244,7 @@ export const TopLevelCollection:TreeNodeCollection =
         key: "religious-other-deities",
         label: "Religious: Other Deities, etc",
         description:'Shankar Parvati, Vishnu. Other drawings in this genre',
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['superuser','sanatan']
     }
     ,
@@ -218,7 +253,7 @@ export const TopLevelCollection:TreeNodeCollection =
         label: "Religious: Gurus And Mystics",
         description:`Gurus and Preachers related to Spirituality, like Lord Dattatreya and Avatars, the Sikh Gurus. 
                         Shirdi Sai sketches being many are under a separate bucket`,
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['superuser','guru']
     }
     
@@ -227,7 +262,7 @@ export const TopLevelCollection:TreeNodeCollection =
         key: "shirdi-sai",
         label: "Religious: Shirdi Sai Baba",
         description:`Sadguru Shirdi Sai I have made so many drawings that it makes logical sense to put this in a separate bucket`,
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['superuser','guru']
     }, {
         key: "otherbeings-places-events",
@@ -236,28 +271,22 @@ export const TopLevelCollection:TreeNodeCollection =
         description:`Pictures of the living - famous human personalities, other living beings (as in May 2024 occasional). Places and famous structures. Events, Misc. Themes. 
                   (But Mumbai City related drawing warrant a separate, independent group!!)`,
         roleBasedDescription:`ADD:Also some chosen religious and spirtual themes: NOTS superuser, sanatan,guru'`,
-        /*
-         if exists rolebaseddescr
-            split by ':'
-            if user roles match third segment NOTS = not or nothing means match the role
-            1st segment, ADD or no spec = add, INSERT = add at begining, REPLACE = replace
-        -- Rest logic here, else if will get too complicated
-        */
-        isLeaf:false,
+       
+        isLeafParent:false,
         accessible:['superuser','people'],// ??=>, 'all']
     }, {
         key: "mumbai-meri-jaan",
         label: "Salaam Mumbai",
 
         description:`Dedication to the City that is part of me that I am part of. Wherever I am`,
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['superuser','people'],// ??=>, 'all']
     },
     {
         key: "transport-and-machines",
         label: "Transport and Machines",
         description:`Trains, Cars, Planes, Ships, Military equipment. Anything that is mechanical`,
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['all']
     },
     ]
@@ -268,7 +297,7 @@ export const ShreeGaneshCollection:TreeNodeCollection =
     parentKey: 'religious-shree-ganesh',
     breadCrumb:[
         {link: '/view', params:'top-level', label: 'HOME'},
-        {link: '/view', params:'religious-shree-ganesh', label: 'Shree Ganesh'}
+        {link: '', params:'religious-shree-ganesh', label: 'Shree Ganesh'}
     ],
     
     levelIndex: 1,
@@ -276,14 +305,14 @@ export const ShreeGaneshCollection:TreeNodeCollection =
         key: "shree-ganesh-b4-q4-2021",
         label: "Shree Ganesh Before Q4 2021",
         description:'Extracted works from the entire collection like best attempts, something new, latest sketches, etc',
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['all']
     },
     {
         key: "shree-ganesh-gte-q4-2021",
         label: "Shree Ganesh  Q4 2021 to Q1 2023",
         description:'Sketches of, and including Shree Ganesh',
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['superuser','sanatan']
     }
     ,
@@ -291,8 +320,8 @@ export const ShreeGaneshCollection:TreeNodeCollection =
         key: "shree-ganesh-gte-q1-2023",
         label: "Shree Ganesh Q1 2023 to Q1 2024",
         description:'Shankar Parvati, Vishnu. Other drawings in this genre',
-        isLeaf:false,
+        isLeafParent:false,
         accessible:['superuser','sanatan']
     }
     ]
-};
+};*/